This has far and away been the most frustrating commit so far
So 1) I am going to hunt down whoever made C++ templates the way they are, and 2) Possibly the only thing worse than templates is std140, because with this I was informed of the rules ahead of time and just forgot that I had to follow them.               Anyway, as for what this commit actually does: pointLights is now transmitted via UBO, and UBOs in general have a spiffy, easy-to-use interface for sending data to the GPU. This was not worth the effort I've put into it.